+++ tests/test_institution.rb
@@ -1,8 +1,8 @@
 require 'test/unit'
 require 'dotenv/load'
 
-require 'nordigen-ruby'
-require 'nordigen_ruby/api/institutions'
+require_relative '../lib/nordigen-ruby'
+require_relative '../lib/nordigen_ruby/api/institutions'
 
 module Nordigen
 
@@ -18,14 +18,14 @@ module Nordigen
         def test_get_institutions
             # Test get list of institutions
             institutions = @institution.get_institutions("LV")
-            id = institutions.each{ |k,v| return k.id if k.id == @institution_id}
+            id = institutions.each{ |k,v| return k["id"] if k["id"] == @institution_id}
             assert_equal(id, @institution_id)
         end
 
         def test_get_institution_by_id
             # Test get institution by id
             institutions = @institution.get_institution_by_id(id=@institution_id)
-            assert_equal(institutions.id, @institution_id)
+            assert_equal(institutions["id"], @institution_id)
         end
 
     end