+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
-## [1.0.0] - 2022-01-13
+## [2.0.0] - 2022-05-05
 
-- Initial release
+- Use JSON response instead of OpenStruct
+- Add country filter
+
 
 ## [1.0.1] - 2022-01-13
 
 - Republish library with latest working gem
+
+## [1.0.0] - 2022-01-13
+
+- Initial release